mkdir model, store and constant and move files
