+++ src/constant/loop_mode.js
